requirements.txt
python-telegram-bot==21.*
openai==1.*
python-dotenv==1.*
